Allow _NET_ACTIVE_WINDOW requests from clients, if they have recent user activity. Rationale:
I don't like allowing clients to activate windows, because they
usually seem to do it for annoying reasons. However, system tray
icons usually activate their main windows when clicked, which is
what the user expects. So, as a comprimise, only allow the action if
the user has recently interacted with them.