[python] Don't rely on Python's GC to close 9P file
There are some 9P files within wmii that takes action based on the
clunk command which is to release the FID and say that we are done
with the file. Examples of these are the /rules and /colrules. That
is, when wmii receive the clunk command for these files it will apply
the rules as the current ones.

In pyxp the File object will close and cleanup either explicitly with
the close function or implictly with the __del__ function that is
triggered from Python's internal garbage collection.

It has been observed that GC is not always triggered after wmiirc
initialization which does not guarantee that any custom rules
written to /colrules for example are actually applied in wmii.

The solution is to explicitly close the File object for Client level
file operations.