Refactor the divider handling from previous improvements
The divider improvements introduced in commit 37f13ec4a still kept
"creating" the dividers on the screen edge even if they never was
grabable which introduced problems when trying to fix the issues in
commits 53d53f72 and 538e5bcf. Since the dividers are kept in a global
structure and are reused for all views (tags) they became grabable again
as the fixes mapped them depending on the focus of floating windows.

The current refactoring will keep track of which dividers are grabable
in the current view (tag) in order to map/unmap depending on the focus
of floating windows. It will also only create the necessary grabable
dividers between the columns on each screen.